fix: backspace issue with gurmukhi keyboard
Refs: #1470
