+++ www/main/navigator/search/components/GurmukhiKeyboard/GurmukhiKeyboard.jsx
@@ -4,7 +4,7 @@ import { getKeyboardKeyValue, getMatraAkhar } from './utils';
 import { Arrow, Spacebar } from './icons';
 import { defaultMatraValue, matras, withMatra, withoutMatra } from './constants';
 
-export const GurmukhiKeyboard = ({ value, searchType, setValue }) => {
+export const GurmukhiKeyboard = ({ query, searchType, setQuery }) => {
   const defaultMatraKeys = Object.keys(defaultMatraValue);
   // If searchType is 2 searchOption is full word
   const isWithMatras = searchType === 2;
@@ -12,25 +12,25 @@ export const GurmukhiKeyboard = ({ value, searchType, setValue }) => {
   const keyboardGrid = [keys];
 
   const handleClick = keyValue => {
-    const lastChar = value.slice(-1);
+    const lastChar = query.slice(-1);
 
     switch (keyValue) {
       case 'meta':
-        if (value !== '') {
-          setValue(value.slice(0, -1));
+        if (query !== '') {
+          setQuery(query.slice(0, -1));
         }
         break;
 
       case 'space':
-        setValue(`${value} `);
+        setQuery(`${query} `);
         break;
 
       default:
         // checks if matra could be applied to last character in searchQuery
         if (!matras.includes(lastChar) && keyValue.includes(lastChar) && keyValue !== lastChar) {
-          setValue(`${value.slice(0, -1)}${keyValue}`);
+          setQuery(`${query.slice(0, -1)}${keyValue}`);
         } else {
-          setValue(value + keyValue);
+          setQuery(query + keyValue);
         }
         break;
     }
@@ -67,12 +67,12 @@ export const GurmukhiKeyboard = ({ value, searchType, setValue }) => {
                       <button
                         type="button"
                         key={i}
-                        data-value={getKeyboardKeyValue(keyboardKey, value)}
+                        data-value={getKeyboardKeyValue(keyboardKey, query)}
                         className={isCurrentKeyDefaultMatraKey ? 'matra-button' : ''}
-                        onClick={() => handleClick(getKeyboardKeyValue(keyboardKey, value))}
+                        onClick={() => handleClick(getKeyboardKeyValue(keyboardKey, query))}
                       >
                         {isCurrentKeyDefaultMatraKey
-                          ? getMatraAkhar(keyboardKey, value)
+                          ? getMatraAkhar(keyboardKey, query)
                           : keyboardKey}
                       </button>
                     );
@@ -88,7 +88,7 @@ export const GurmukhiKeyboard = ({ value, searchType, setValue }) => {
 };
 
 GurmukhiKeyboard.propTypes = {
-  value: PropTypes.string,
-  setValue: PropTypes.func,
+  query: PropTypes.string,
+  setQuery: PropTypes.func,
   searchType: PropTypes.number,
 };
